Replace locked accounts by locked balances
"spotted a bug in the way locked accounts are handled. If Gizer issues tokens to team and advisors, the tokens in these accounts gets locked during a lockup period. If any of these accounts contributes to the ICO, the tokens issued also remain locked." Rather than lock accounts, we now track the amount of locked tokens in each account.